AQI_monitoring/Air_Quality_Monitoring_System/requirements.txt
fastapi>=0.100.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.11.0
python-multipart>=0.0.5
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
email-validator>=2.0.0
uvicorn>=0.22.0
python-dotenv>=0.21.0
psycopg2-binary>=2.9.1
APScheduler>=3.10.0